A couple of changes to the Java interface:   - Variable's id are of long type (like space dimension arguments);   - added NOT_EQUAL value to enumeration Relation_Symbol.

+++ interfaces/Java/parma_polyhedra_library/Linear_Expression_Variable.java
@@ -45,7 +45,7 @@ extends Linear_Expression {
 	return new Linear_Expression_Variable(arg);
     }
 
-    private int var_id() {
+    private long var_id() {
         return arg.id();
     }
